+++ src/benches.rs
@@ -152,7 +152,7 @@ fn epoll_wait(b: &mut Bencher) {
 fn handle_interface_data(b: &mut Bencher) {
     let mut node = GenericCloud::<Frame>::new(
         MAGIC, Device::dummy("vpncloud0", "/dev/null", Type::Tap).unwrap(), 0,
-        Box::new(SwitchTable::new(300)), 1800, true, true, vec![], Crypto::None
+        Box::new(SwitchTable::new(300)), 1800, true, true, vec![], Crypto::None, None
     );
     let mut data = [0; 1500];
     data[105] = 45;
@@ -166,7 +166,7 @@ fn handle_interface_data(b: &mut Bencher) {
 fn handle_net_message(b: &mut Bencher) {
     let mut node = GenericCloud::<Frame>::new(
         MAGIC, Device::dummy("vpncloud0", "/dev/null", Type::Tap).unwrap(), 0,
-        Box::new(SwitchTable::new(300)), 1800, true, true, vec![], Crypto::None
+        Box::new(SwitchTable::new(300)), 1800, true, true, vec![], Crypto::None, None
     );
     let addr = SocketAddr::V4(SocketAddrV4::new(Ipv4Addr::new(127, 0, 0, 1), 1));
     let mut data = [0; 1500];